+++ pesu-components/src/hooks/useAutoFocus/index.ts
@@ -1,0 +1,1 @@
+export { default as useAutoFocus } from './useAutoFocus';
